move serialize_niimg to _utils.testing
